Medication feature enhance
