+++ lib/core/services/medication_service.dart
@@ -1,13 +1,17 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:nursery_app/core/services/classroom_teacher_service.dart';
+import 'package:nursery_app/core/services/student_parent_service.dart';
 import 'package:nursery_app/models/medications/medication_model.dart';
 
 class MedicationService {
   final FirebaseFirestore _firestore = FirebaseFirestore.instance;
-  final String _collectionName = 'medications';
+  final String _db = 'medications';
+  final ClassroomTeacherService _classroomTeacherService =
+      ClassroomTeacherService();
 
   // Fetch all medications
   Stream<List<Medication>> getAllMedications() {
-    return _firestore.collection(_collectionName).snapshots().map((snapshot) {
+    return _firestore.collection(_db).snapshots().map((snapshot) {
       return snapshot.docs.map((doc) {
         return Medication.fromFirestore(doc.id, doc.data());
       }).toList();
@@ -15,10 +19,10 @@ class MedicationService {
   }
 
   // Fetch medications for a specific child
-  Stream<List<Medication>> getMedicationsForChild(String childId) {
+  Stream<List<Medication>> getMedicationsForChild(String studentId) {
     return _firestore
-        .collection(_collectionName)
-        .where('childId', isEqualTo: childId)
+        .collection(_db)
+        .where('studentId', isEqualTo: studentId)
         .snapshots()
         .map((snapshot) {
       return snapshot.docs.map((doc) {
@@ -28,10 +32,10 @@ class MedicationService {
   }
 
   // Fetch medications for a parent (all their children)
-  Stream<List<Medication>> getMedicationsForParent(List<String> childIds) {
+  Stream<List<Medication>> getMedicationsForParent(String parentId) {
     return _firestore
-        .collection(_collectionName)
-        .where('childId', whereIn: childIds)
+        .collection(_db)
+        .where('reportedByUserId', isEqualTo: parentId)
         .snapshots()
         .map((snapshot) {
       return snapshot.docs.map((doc) {
@@ -43,8 +47,7 @@ class MedicationService {
   // Fetch a single medication
   Future<Medication?> getMedication(String medicationId) async {
     try {
-      final doc =
-          await _firestore.collection(_collectionName).doc(medicationId).get();
+      final doc = await _firestore.collection(_db).doc(medicationId).get();
       if (doc.exists) {
         return Medication.fromFirestore(doc.id, doc.data()!);
       }
@@ -52,6 +55,47 @@ class MedicationService {
     } catch (e) {
       throw Exception('Failed to fetch medication: $e');
     }
+  }
+
+  // Fetch medications for a teacher
+  Stream<List<Medication>> getMedicationsForTeacher(String teacherId) async* {
+    // Get classroom IDs where teacherId matches
+    final classroomIds = await _firestore
+        .collection('classroomTeachers')
+        .where('teacherId', isEqualTo: teacherId)
+        .get()
+        .then((snapshot) => snapshot.docs
+            .map((doc) => doc.data()['classroomId'] as String)
+            .toList());
+    if (classroomIds.isEmpty) {
+      yield* Stream.value([]);
+      return;
+    }
+
+    // Get student IDs from those classrooms
+    List<String> studentIds = [];
+    for (var classroomId in classroomIds) {
+      final classroomStudentIds = await _firestore
+          .collection('students')
+          .where('classroomId', isEqualTo: classroomId)
+          .get()
+          .then((snapshot) => snapshot.docs.map((doc) => doc.id).toList());
+      studentIds.addAll(classroomStudentIds);
+    }
+    if (studentIds.isEmpty) {
+      yield* Stream.value([]);
+      return;
+    }
+    // Get medications for those students
+    yield* _firestore
+        .collection(_db)
+        .where('studentId', whereIn: studentIds)
+        .snapshots()
+        .map((snapshot) {
+      return snapshot.docs.map((doc) {
+        return Medication.fromFirestore(doc.id, doc.data());
+      }).toList();
+    });
   }
 
   // Update an existing medication
@@ -65,7 +109,7 @@ class MedicationService {
     required String photoUrl,
   }) async {
     try {
-      await _firestore.collection(_collectionName).doc(medicationId).update({
+      await _firestore.collection(_db).doc(medicationId).update({
         'medicationName': medicationName,
         'dosage': dosage,
         'frequency': frequency,
@@ -82,7 +126,7 @@ class MedicationService {
   // Delete a medication
   Future<void> deleteMedication(String medicationId) async {
     try {
-      await _firestore.collection(_collectionName).doc(medicationId).delete();
+      await _firestore.collection(_db).doc(medicationId).delete();
     } catch (e) {
       throw Exception('Failed to delete medication: $e');
     }
@@ -90,7 +134,7 @@ class MedicationService {
 
   // Create a new medication
   Future<void> createMedication({
-    required String childId,
+    required String studentId,
     required String medicationName,
     required String dosage,
     required String frequency,
@@ -99,11 +143,11 @@ class MedicationService {
     required String photoUrl,
   }) async {
     try {
-      final docRef = _firestore.collection(_collectionName).doc();
+      final docRef = _firestore.collection(_db).doc();
 
       final medication = Medication(
         id: docRef.id,
-        childId: childId,
+        studentId: studentId,
         medicationName: medicationName,
         dosage: dosage,
         frequency: frequency,
@@ -144,10 +188,7 @@ class MedicationService {
         ]);
       }
 
-      await _firestore
-          .collection(_collectionName)
-          .doc(medicationId)
-          .update(updates);
+      await _firestore.collection(_db).doc(medicationId).update(updates);
     } catch (e) {
       throw Exception('Failed to update medication: $e');
     }
